# front/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Le Renard Étudiant — API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/comptes/inscription": {
      "post": {
        "tags": [
          "comptes"
        ],
        "summary": "Inscription",
        "description": "Crée un compte et envoie l'email de confirmation.\n\nSi l'email a déjà un compte, la réponse est la même : on ne révèle pas qui est inscrit.\nLe titulaire de l'adresse reçoit un email qui le lui signale.\n\nSans vérification de l'email (ADR 0023), le compte est utilisable tout de suite,\net une adresse déjà inscrite est signalée : personne d'autre ne pourrait prévenir son titulaire.",
        "operationId": "inscription_api_comptes_inscription_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inscription"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/comptes/confirmation": {
      "post": {
        "tags": [
          "comptes"
        ],
        "summary": "Confirmation",
        "description": "Confirme l'adresse d'un nouveau compte, ou la nouvelle adresse d'un compte existant.",
        "operationId": "confirmation_api_comptes_confirmation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JetonRecu"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/comptes/renvoyer-confirmation": {
      "post": {
        "tags": [
          "comptes"
        ],
        "summary": "Renvoyer Confirmation",
        "operationId": "renvoyer_confirmation_api_comptes_renvoyer_confirmation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandeParEmail"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/comptes/connexion": {
      "post": {
        "tags": [
          "comptes"
        ],
        "summary": "Connexion",
        "operationId": "connexion_api_comptes_connexion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Connexion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Moi"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comptes/moi": {
      "get": {
        "tags": [
          "comptes"
        ],
        "summary": "Moi",
        "operationId": "moi_api_comptes_moi_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Moi"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "comptes"
        ],
        "summary": "Modifier Profil",
        "operationId": "modifier_profil_api_comptes_moi_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificationProfil"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Moi"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/comptes/moi/mot-de-passe": {
      "post": {
        "tags": [
          "comptes"
        ],
        "summary": "Changer Mot De Passe",
        "description": "Change le mot de passe et déconnecte les autres appareils.",
        "operationId": "changer_mot_de_passe_api_comptes_moi_mot_de_passe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangementMotDePasse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/comptes/moi/email": {
      "post": {
        "tags": [
          "comptes"
        ],
        "summary": "Changer Email",
        "description": "Envoie un lien à la nouvelle adresse. Elle ne remplace l'ancienne qu'après confirmation.\n\nSans vérification de l'email (ADR 0023), la nouvelle adresse remplace l'ancienne tout de suite.",
        "operationId": "changer_email_api_comptes_moi_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangementEmail"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/comptes/mot-de-passe-oublie": {
      "post": {
        "tags": [
          "comptes"
        ],
        "summary": "Mot De Passe Oublie",
        "description": "Envoie un lien de réinitialisation. Même réponse si l'email est inconnu.",
        "operationId": "mot_de_passe_oublie_api_comptes_mot_de_passe_oublie_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemandeParEmail"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/comptes/reinitialisation": {
      "post": {
        "tags": [
          "comptes"
        ],
        "summary": "Reinitialisation",
        "description": "Change le mot de passe et ferme toutes les sessions.",
        "operationId": "reinitialisation_api_comptes_reinitialisation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reinitialisation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangementEmail": {
        "properties": {
          "nouvel_email": {
            "type": "string",
            "format": "email",
            "title": "Nouvel Email"
          },
          "mot_de_passe": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Mot De Passe"
          }
        },
        "type": "object",
        "required": [
          "nouvel_email",
          "mot_de_passe"
        ],
        "title": "ChangementEmail"
      },
      "ChangementMotDePasse": {
        "properties": {
          "actuel": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Actuel"
          },
          "nouveau": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Nouveau"
          }
        },
        "type": "object",
        "required": [
          "actuel",
          "nouveau"
        ],
        "title": "ChangementMotDePasse"
      },
...
      "Connexion": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "mot_de_passe": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Mot De Passe"
          }
        },
        "type": "object",
        "required": [
          "email",
          "mot_de_passe"
        ],
        "title": "Connexion"
      },
...
      "DemandeParEmail": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "DemandeParEmail"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Inscription": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "pseudo": {
            "type": "string",
            "maxLength": 30,
            "minLength": 3,
            "pattern": "^[\\w.-]+$",
            "title": "Pseudo"
          },
          "mot_de_passe": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Mot De Passe"
          }
        },
        "type": "object",
        "required": [
          "email",
          "pseudo",
          "mot_de_passe"
        ],
        "title": "Inscription"
      },
      "JetonRecu": {
        "properties": {
          "jeton": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Jeton"
          }
        },
        "type": "object",
        "required": [
          "jeton"
        ],
        "title": "JetonRecu"
      },
...
      "ModificationProfil": {
        "properties": {
          "pseudo": {
            "type": "string",
            "maxLength": 30,
            "minLength": 3,
            "pattern": "^[\\w.-]+$",
            "title": "Pseudo"
          }
        },
        "type": "object",
        "required": [
          "pseudo"
        ],
        "title": "ModificationProfil"
      },
...
      "Moi": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "pseudo": {
            "type": "string",
            "title": "Pseudo"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "licence_acceptee": {
            "type": "boolean",
            "title": "Licence Acceptee"
          },
          "jeton_csrf": {
            "type": "string",
            "title": "Jeton Csrf"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "pseudo",
          "role",
          "licence_acceptee",
          "jeton_csrf"
        ],
        "title": "Moi",
        "description": "L'utilisateur connecté. Le jeton CSRF est à renvoyer dans l'en-tête X-CSRF-Token."
      },
...
      "Reinitialisation": {
        "properties": {
          "jeton": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Jeton"
          },
          "nouveau": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "title": "Nouveau"
          }
        },
        "type": "object",
        "required": [
          "jeton",
          "nouveau"
        ],
        "title": "Reinitialisation"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "apprenant",
          "contributeur",
          "relecteur",
          "admin"
        ],
        "title": "Role",
        "description": "Rôles du cadrage, du moins au plus élevé. Chaque rôle a les droits des précédents."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}